AoC 2015 day 15
Used a greedy solution in part 1, always adding the ingredient which maximizes the score at that time. Pretty sure that's generally optimal.

Took a brute force approach for part 2. There are only 4 ingredients, so there are 103 choose 3 possible recipes, which comfortably fits in memory. Filter for those with the correct amount of calories, then find the highest scoring recipe and done.