AoC 2024 Day 14
Easy part 1, but I struggled to find a heuristic for part 2. I eventually settled on checking standard deviation along the axes in the hope that things were clustering, which wound up reducing the search space enough I could click through manually.

I added a check for an NxN block after finding the answer to speed things up and remove the manual element. Still not sure what the best way to go about this was.