AoC 2024 Day 21 - part 2
Oof. Tracking pairs of sequential characters got me to the finish line. But an optimistic assumption I made in part 1 bit me -- it turns out that the paths which are shortest for a single expansion are not optimal after further expansions.

I'd originally removed such paths from my cache, but had to add them back to get this working. The code is therefore quite slow for part 2, but hey, it works.