AoC 2024 Day 24
Umm that was rough. I wound up writing a recursive approach with a set of hardcoded tests, evaluating up to a point to verify correctness, then generating and applying candidate swaps when expected and actual diverged.

Then spent a long time debugging. It turns out that there were some swaps I could do which produce the right output for my input and the test values, but are not correct in the general case. I had to manually inspect the graph to eliminate the candidate wire switch which caused problems.